decouple theme detection
