+++ src/components/conversation_page/WatchAndStudyTab.jsx
@@ -3,47 +3,18 @@ import { defaultLayoutIcons, DefaultVideoLayout } from "@vidstack/react/player/l
 import "@vidstack/react/player/styles/default/layouts/video.css";
 import "@vidstack/react/player/styles/default/theme.css";
 import PropTypes from "prop-types";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { IoInformationCircleOutline } from "react-icons/io5";
 import { isElectron } from "../../utils/isElectron";
-import { useTheme } from "../../utils/ThemeContext/useTheme";
+import { useAutoDetectTheme } from "../../utils/ThemeContext/useAutoDetectTheme";
 
 const WatchAndStudyTab = ({ videoUrl, subtitleUrl, dialog, skillCheckmark }) => {
     const { t } = useTranslation();
     const [highlightState, setHighlightState] = useState({});
     const [iframeLoading, setiFrameLoading] = useState(true);
 
-    const { theme } = useTheme();
-    const [, setCurrentTheme] = useState(theme);
-    const [isDarkMode, setIsDarkMode] = useState(false);
-
-    useEffect(() => {
-        const mediaQuery = window.matchMedia("(prefers-color-scheme: dark)");
-
-        const updateTheme = () => {
-            if (theme === "auto") {
-                const systemPrefersDark = mediaQuery.matches;
-                setCurrentTheme(systemPrefersDark ? "dark" : "light");
-                setIsDarkMode(systemPrefersDark);
-            } else {
-                setCurrentTheme(theme);
-                setIsDarkMode(theme === "dark");
-            }
-        };
-
-        // Initial check and listener
-        updateTheme();
-        if (theme === "auto") {
-            mediaQuery.addEventListener("change", updateTheme);
-        }
-
-        return () => {
-            mediaQuery.removeEventListener("change", updateTheme);
-        };
-    }, [theme]);
-
-    const videoColorScheme = isDarkMode ? "dark" : "light";
+    const { autoDetectedTheme } = useAutoDetectTheme();
 
     const handleIframeLoad = () => setiFrameLoading(false);
 
@@ -59,8 +30,8 @@ const WatchAndStudyTab = ({ videoUrl, subtitleUrl, dialog, skillCheckmark }) => 
         <div className="grid grid-cols-1 gap-6 md:grid-cols-2">
             <div>
                 <div className="text-xl font-semibold">{t("tabConversationExam.watchCard")}</div>
-                <div className="divider divider-secondary mb-4 mt-0"></div>
-                <div className="top-[calc(14rem)] bg-base-100 md:sticky md:z-10">
+                <div className="divider divider-secondary mt-0 mb-4"></div>
+                <div className="bg-base-100 top-[calc(14rem)] md:sticky md:z-10">
                     <div className="aspect-video">
                         <div className="relative h-full w-full">
                             {isElectron() && videoUrl && videoUrl.startsWith("http://localhost") ? (
@@ -68,7 +39,7 @@ const WatchAndStudyTab = ({ videoUrl, subtitleUrl, dialog, skillCheckmark }) => 
                                     <MediaProvider />
                                     <DefaultVideoLayout
                                         icons={defaultLayoutIcons}
-                                        colorScheme={videoColorScheme}
+                                        colorScheme={autoDetectedTheme}
                                     />
                                     <Track
                                         src={subtitleUrl}
@@ -110,8 +81,8 @@ const WatchAndStudyTab = ({ videoUrl, subtitleUrl, dialog, skillCheckmark }) => 
             </div>
             <div>
                 <div className="text-xl font-semibold">{t("tabConversationExam.studyCard")}</div>
-                <div className="divider divider-secondary mb-4 mt-0"></div>
-                <div className="collapse collapse-arrow bg-base-200 dark:bg-slate-700">
+                <div className="divider divider-secondary mt-0 mb-4"></div>
+                <div className="collapse-arrow bg-base-200 collapse dark:bg-slate-700">
                     <input type="checkbox" />
                     <button type="button" className="collapse-title text-start font-semibold">
                         {t("tabConversationExam.studyExpandBtn")}
