+++ src/components/sound_page/hooks/useSoundVideoDialog.jsx
@@ -1,9 +1,9 @@
 import { MediaPlayer, MediaProvider } from "@vidstack/react";
 import { defaultLayoutIcons, DefaultVideoLayout } from "@vidstack/react/player/layouts/default";
 import PropTypes from "prop-types";
-import { useEffect, useRef, useState } from "react";
+import { useRef, useState } from "react";
 import { IoInformationCircleOutline } from "react-icons/io5";
-import { useTheme } from "../../../utils/ThemeContext/useTheme";
+import { useAutoDetectTheme } from "../../../utils/ThemeContext/useAutoDetectTheme";
 import { SoundVideoDialogContext } from "./useSoundVideoDialogContext";
 
 export const SoundVideoDialogProvider = ({ children, t }) => {
@@ -24,26 +24,7 @@ export const SoundVideoDialogProvider = ({ children, t }) => {
 
     const dialogRef = useRef(null);
     const mediaPlayerRef = useRef(null);
-    const { theme } = useTheme();
-    const [isDarkMode, setIsDarkMode] = useState(false);
-
-    useEffect(() => {
-        const mediaQuery = window.matchMedia("(prefers-color-scheme: dark)");
-
-        const updateTheme = () => {
-            const systemPrefersDark = mediaQuery.matches;
-            setIsDarkMode(theme === "auto" ? systemPrefersDark : theme === "dark");
-        };
-
-        updateTheme();
-        if (theme === "auto") {
-            mediaQuery.addEventListener("change", updateTheme);
-        }
-
-        return () => {
-            mediaQuery.removeEventListener("change", updateTheme);
-        };
-    }, [theme]);
+    const { autoDetectedTheme } = useAutoDetectTheme();
 
     const showDialog = (state) => {
         setDialogState({ ...state, isOpen: true, iframeLoading: true });
@@ -100,7 +81,7 @@ export const SoundVideoDialogProvider = ({ children, t }) => {
                                         <MediaProvider />
                                         <DefaultVideoLayout
                                             icons={defaultLayoutIcons}
-                                            colorScheme={isDarkMode ? "dark" : "light"}
+                                            colorScheme={autoDetectedTheme}
                                         />
                                     </MediaPlayer>
                                 ) : dialogState.isOpen && dialogState.videoUrl ? (
