+++ src/components/word_page/useWaveformTheme.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from "react";
+import { useEffect, useState } from "react";
 
 const useWaveformTheme = (
     theme,
